delete kubeconfig from yaml. move to db (#61)
* delete kubeconfig from yaml. move to db

* delete fsconfig

* merge PV/PVC config into server config
