+++ pkg/common/config/server.go
@@ -17,6 +17,11 @@ limitations under the License.
 package config
 
 import (
+	"os"
+
+	apiv1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/yaml"
+
 	"paddleflow/pkg/common/logger"
 	"paddleflow/pkg/common/schema"
 )
@@ -24,11 +29,16 @@ import (
 var DefaultRunYamlPath string = "./run.yaml"
 var serverDefaultConfPath = "./config/server/default/paddleserver.yaml"
 
+// DefaultPV the global default pv instance
+var DefaultPV *apiv1.PersistentVolume
+
+// DefaultPVC the global default pvc instance
+var DefaultPVC *apiv1.PersistentVolumeClaim
+
 type ServerConfig struct {
 	Database      DatabaseConfig            `yaml:"database"`
 	Log           logger.LogConfig          `yaml:"log"`
 	ApiServer     ApiServerConfig           `yaml:"apiServer"`
-	KubeConfig    KubeConfig                `yaml:"kubeConfig"`
 	Job           JobConfig                 `yaml:"job"`
 	Fs            FsServerConf              `yaml:"fs"`
 	NamespaceList []string                  `yaml:"namespaceList"`
@@ -103,4 +113,24 @@ var (
 
 func InitConfigFromDefaultYaml(conf interface{}) error {
 	return InitConfigFromYaml(conf, serverDefaultConfPath)
+}
+
+// InitDefaultPV initialize the default pv instance
+func InitDefaultPV(path string) error {
+	reader, err := os.Open(path)
+	if err != nil {
+		return err
+	}
+	defer reader.Close()
+	return yaml.NewYAMLOrJSONDecoder(reader, 1024).Decode(&DefaultPV)
+}
+
+// InitDefaultPVC initialize the default pvc instance
+func InitDefaultPVC(path string) error {
+	reader, err := os.Open(path)
+	if err != nil {
+		return err
+	}
+	defer reader.Close()
+	return yaml.NewYAMLOrJSONDecoder(reader, 1024).Decode(&DefaultPVC)
 }
